Drop maze evolution reproducibility assertions — evolution is stochastic
Replace golden/dual-run equality checks with a subprocess isolation test
that only verifies finite, internally consistent champion metrics.
